REBASING FIX - create new board file
So the head of the old birdwing linux tree wasn't quite a clean 3.8
davinci kernel.  I tried to find all changes by grepping around for
"manhattan" and replaying them.  The changes for this commit are:

Copy the omap138 hawkboard file as a new board file, edit some configs
to allow selection of the new board file.