Add driver for OSD touchscreen
This is a (heavily) modified version of the touchscreen drivers given us
by OSD. They have been modernized a bit and also made to conform to the
datasheet they gave us, which wasn't true before.

It is an i2c device that sits on top of, right now, the i2c0 adapter. It
registers as a touchscreen and uses the multitouch libraries to report
output. Right now, it lives on /dev/input/event2.

It is still very raw and doesn't have good settings for tuning what does
and doesn't count as a touch; it doesn't report pressure; and it doesn't
do anything about large object rejection.