Add some new wireless drivers
I added bridging support since connman's wifi AP support is most easily
configured for tethering, and a lot of bluetooth drivers, and some other
things that looked useful.  I also reconfigured a lot of drivers from
being modules to being compiled in, and I removed a few drivers that
looked particularly useless.  And then I ran the config through
savedefconfig, because checking in a full config as a defconfig is
stupid.