Run the checked in defconfig through savedefconfig
The checked in defconfig is actually a full config, which is not
supposed to be editted by hand.  By attempting to edit it by hand
we are obscuring what the actual kernel configuration is, and
giving the impression that anything we add to this file will
automatically be defined when compiling the kernel.

I have verified that this defconfig produces an identical .config to
the "defconfig" from the parent commit.