Prevent SPI from leaving CS hanging
Here are some facts:
- Calls to the sleep() family, like nsleep() or usleep() are allowed to
sleep your thread for longer than you specify.
- Sometimes, they will sleep your thread for _lots_ longer than you
specify.
- The spi-bitbang base driver completes an ongoing transaction, _then_
sleeps, _then_ deasserts chipselect.
- Consequently, very occasionally a transaction will mark itself as
completed and return control to userspace, which queues another
transaction, but the spi_bitbang workqueue is sleeping in the
background. When it wakes up, it deasserts chipselect and continues with
its work. The problem is, that means chipselect is only deasserted for a
few tens of microseconds. And that just won't do.